Moved the selection value exporter from databrowser2 to the databrowser2.opiwidget
