+++ applications/features/org.csstudio.trends.databrowser2.feature/feature.xml
@@ -66,7 +66,6 @@ For details, see http://www.eclipse.org/legal/epl-v10.html
       <import plugin="org.eclipse.ui.ide" version="3.6.0" match="greaterOrEqual"/>
       <import plugin="org.csstudio.trends.databrowser2" version="3.2.0" match="greaterOrEqual"/>
       <import plugin="org.csstudio.swt.xygraph" version="1.1.3" match="greaterOrEqual"/>
-      <import plugin="org.csstudio.utility.pvmanager.extra" version="3.2.0" match="greaterOrEqual"/>
    </requires>
 
    <plugin
