+++ applications/plugins/org.csstudio.trends.databrowser2/META-INF/MANIFEST.MF
@@ -34,8 +34,7 @@ Require-Bundle: org.csstudio.utility.test;resolution:=optional,
  org.csstudio.utility.pvmanager.vtype;bundle-version="2.0.0",
  org.csstudio.utility.pvmanager.loc;bundle-version="3.1.0";resolution:=optional,
  org.csstudio.utility.pvmanager.sim;bundle-version="3.1.0";resolution:=optional,
- org.csstudio.autocomplete.ui,
- org.csstudio.utility.pvmanager.extra;bundle-version="3.2.0"
+ org.csstudio.autocomplete.ui
 Bundle-ActivationPolicy: lazy
 Bundle-RequiredExecutionEnvironment: JavaSE-1.7
 Bundle-Vendor: Kay Kasemir <kasemirk@ornl.gov> - SNS
